HW7: Swagger

+++ MetricsAgentsManager/DAL/Interfaces/IAgentsRepository.cs
@@ -8,10 +8,10 @@ namespace MetricsAgentsManager.DAL.Interfaces
     public interface IAgentsRepository : IRepository<AgentInfo>
     {
         IList<AgentInfo> GetActiveAgents();
-        IList<CpuMetric> GetCpuMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, int agentId = 0);
-        IList<NetworkMetric> GetNetworkMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, int agentId = 0);
-        IList<DotNetMetric> GetDotNetMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, int agentId = 0);
-        IList<HddSpaceMetric> GetHddSpaceMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, int agentId = 0);
-        IList<RamMetric> GetRamMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, int agentId = 0);
+        IList<CpuMetric> GetCpuMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, long agentId = 0);
+        IList<NetworkMetric> GetNetworkMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, long agentId = 0);
+        IList<DotNetMetric> GetDotNetMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, long agentId = 0);
+        IList<HddSpaceMetric> GetHddSpaceMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, long agentId = 0);
+        IList<RamMetric> GetRamMetricsByPeriod(DateTimeOffset fromTime, DateTimeOffset toTime, long agentId = 0);
     }
 }
